Move ErrorMessageContent to its own file
