+++ src/injected/frameCommunicators/error-message-content.ts
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+export interface ErrorMessageContent {
+    name: string;
+    message: string;
+    stack: string;
+}
