+++ src/injected/tab-stops-listener.ts
@@ -6,9 +6,9 @@ import { HTMLElementUtils } from '../common/html-element-utils';
 import { DateProvider } from './../common/date-provider';
 import { WindowUtils } from './../common/window-utils';
 import { VisualizationWindowMessage } from './drawing-controller';
+import { ErrorMessageContent } from './frameCommunicators/error-message-content';
 import { FrameCommunicator, IMessageRequest } from './frameCommunicators/frame-communicator';
 import { FrameMessageResponseCallback } from './frameCommunicators/window-message-handler';
-import { ErrorMessageContent } from './frameCommunicators/window-message-marshaller';
 import { ScannerUtils } from './scanner-utils';
 
 export interface TabStopEvent {
